minor renaming; instead of exposing a constructor ShadyRoot, just expose attachShadow as this is more inline with native.
